third_party/blink/web_tests/platform/generic/fast/forms/file/file-input-change-event-expected.txt

 

This tests the condition that triggers a 'change' event on file input forms.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test that the 'change' event is triggered on a single file form when a selected file is changed:
PASS singleFileInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS singleFileInput.value is "C:\\fakepath\\bar.txt"
PASS changeDispatched is true
PASS singleFileInput.value is "C:\\fakepath\\bar.txt"
PASS changeDispatched is false
PASS singleFileInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS singleFileInput.value is "C:\\fakepath\\baz.png"
PASS changeDispatched is true


Test that the 'change' event is triggered on a multiple file form when a selected file is changed:
PASS multipleFilesInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\bar.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\bar.txt"
PASS changeDispatched is false
PASS multipleFilesInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
Test that the 'change' event is triggered on a multiple file form when selected files are changed:
PASS multipleFilesInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\foo.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\bar.txt"
PASS changeDispatched is true
PASS multipleFilesInput.value is "C:\\fakepath\\bar.txt"
PASS changeDispatched is false
PASS successfullyParsed is true

TEST COMPLETE
